requirements.txt
numpy>=1.26
pandas>=2.1
scikit-learn>=1.4
lightgbm>=4.3
xgboost>=2.0
matplotlib>=3.8
pyyaml>=6.0
